Fix a bug that broke global variables in reloaded methods.
Basically the reloaded methods still take the global variable from the newly loaded module, where it was originally from.

The solution is to always keep that in memory, until the next reload. Methods of classes are always reloaded since there is no way to check if they have changed since the last version. So on a new reload the module gets reloaded, the methods replaced, and the methods now point to the newest global variable instance.